+++ Changes.md
@@ -11,6 +11,7 @@ you spot any mistakes.
 * Clone connection config for new pool connections
 * Default `connectTimeout` to 2 minutes
 * Reject unauthorized SSL connections (use `ssl.rejectUnauthorized` to override) #816
+* Return last error when PoolCluster exhausts connection retries #818
 * Throw on unknown SSL profile name #817
 * User newer TLS functions when available #809